Wat aanpassingen
Hij kan nu bewegen(w of s ingedrukt houden werkt alleen niet). Je begint door op 'n' te drukken, dan kun je met w omhoog en met s omlaag. Als je op b drukt verandert het veld in een leeg veld met alleen 1 bullet.
Er zaten nog wat fouten in onze eerste versie die ik er uit heb gehaald. Ik heb dit opgelost dmv newGstate, vgsW en vgsS

+++ src/Player.hs
@@ -6,5 +6,5 @@ type Position = (Float, Float)
 
 type Health = Int
 
-type Speed = Int
+type Speed = Float
